+++ CHANGELOG.md
@@ -85,6 +85,7 @@ The following NPM package is removed:
 ### Fixed
 - Issue #808: Fix geometry polygon-with-hole for searchBar/specialWfs.
 - Issue #813: Fix various WMS-T bugs and styling.
+- Save Selection: takes over the transparency of the layer again.
 
 ---
 ## v2.24.1 - 2022-08-19